Exercise 5 of 9
You should connect to the database named learnyoumongo and insert a
document into the docs collection.

The document should be a json document with the following properties:

  * `firstName`
  * `lastName`

firstName will be passed as the first argument to the lesson.
lastName will be passed as the second argument to the lesson.

Use console.log to print out the object used to create the document.